Tenant-specific user should be able to be substituted by users from the same tenant only: move class 4240
